Interface to mallet topicmodelling.

Functionality to support the following workflow (see examples): (a) Turn partition_bundle-object into mallet instance list, (b) store the resulting jobjRef-object, (c) run mallet topic modelling and (d) turn ParallelTopicModel Java object into LDA_Gibbs object from package topicmodels.

as.instance_list(x, ...)

# S4 method for partition_bundle
as.instance_list(x, p_attribute = "word", verbose = TRUE)

# S4 method for list
as.instance_list(x, corpus, p_attribute = "word")

mallet_instance_list_store(x, filename = tempfile())

mallet_instance_list_load(filename)

Arguments

x
A partition_bundle object.

...
further parameters

p_attribute
The p_attribute to use, typically "word" or "lemma".

verbose
A logical value, whether to be verbose.

corpus
A CWB indexed corpus, defined either by corpus ID, or corpus object.

filename
Where to store the Java-object.

Details

The function mallet_instance_list_load will load a Java InstanceList object that has been saved to disk (e.g. by using the mallet_instance_list_store function). The return value is a jobjRef object. Internally, the function reuses code of the function load.mallet.instances from the R package mallet.

Examples

# Preparations: Create instance list

if (!mallet_is_installed()) mallet_install()

#> ... downloading mallet

#> ... initializing JVM

#> ... adding to classpath

polmineR::use("polmineR")

#> ... activating corpus: GERMAPARLMINI (version: 0.0.1 | build date: 2019-02-23)

#> ... activating corpus: REUTERS

speeches <- polmineR::as.speeches("GERMAPARLMINI", s_attribute_name = "speaker")

  instance_list <- as.instance_list(speeches)

#> ... create alphabet

#> ... instantiate pipe

#> ... decode token stream

#> ... creating vector of document ids

#> ... creating vector of corpus positions

#> ... decoding character vectors

#> ... generating list of character vectors

#> ... created instances

  lda <- rJava::.jnew("cc/mallet/topics/ParallelTopicModel", 25L, 5.1, 0.1)
  lda$addInstances(instance_list)
  lda$setNumThreads(1L)
  lda$setTopicDisplay(50L, 10L)
  lda$setNumIterations(150L)
  lda$estimate()
  # lda$write(rJava::.jnew("java/io/File", destfile))

# Load topicmodel and turn it into LDA_Gibbs

# mallet_lda <- mallet_load_topicmodel(destfile)
topicmodels_lda <- as_LDA(lda)

#> ... getting number of documents and number of terms

#> ... getting alphabet

#> ... getting document names

#> ... getting topic probabilities (gamma matrix)

#> ... getting topic word weights (beta matrix)

library(polmineR)
use("polmineR")

#> ... activating corpus: GERMAPARLMINI (version: 0.0.1 | build date: 2019-02-23)

#> ... activating corpus: REUTERS

speeches <- as.speeches("GERMAPARLMINI", s_attribute_name = "speaker")
speeches_instance_list <- as.instance_list(speeches, p_attribute = "word")

#> ... create alphabet

#> ... instantiate pipe

#> ... decode token stream

#> ... creating vector of document ids

#> ... creating vector of corpus positions

#> ... decoding character vectors

#> ... generating list of character vectors

#> ... created instances

speeches <- as.speeches("GERMAPARLMINI", s_attribute_name = "speaker")
id_list <- p_attributes(speeches, p_attribute = "word", decode = FALSE)

#> ... creating vector of document ids

#> ... creating vector of corpus positions

#> ... decoding character vectors

#> ... generating list of character vectors

instance_list <- as.instance_list(id_list, corpus = "GERMAPARLMINI", p_attribute = "word")